requirements.txt
django
djangorestframework
django-heroku
gunicorn
requests